Iridium RAAN scan for our orbit
implementation of the RAAN scan for the 500km SSO using STELA for the sat long-term, and j2 for the sat short term and the constellation long and short term. Code performs a 24h dt=5s visibility sim for one out of 5 days over a year.
Also fixed GetIridiumPasses needing the t array but not asking for it as an argument